Make sure `AbortableSpawner` can't be cloned
* Leave Spawn traits in `spawner.rs` only
* Add `AbortableSpawnerWeak`
* Rename `CoinFutureSpawner` to `CoinFutSpawner`
* Rename `MmSpawner` to `MmFutSpawner`
