+++ mm2src/coins/eth.rs
@@ -22,7 +22,7 @@
 //
 use async_trait::async_trait;
 use bitcrypto::{keccak256, sha256};
-use common::executor::Timer;
+use common::executor::{AbortableSpawner, Timer};
 use common::log::{error, info, warn};
 use common::{get_utc_timestamp, now_ms, small_rng, DEX_FEE_ADDR_RAW_PUBKEY};
 use crypto::privkey::key_pair_from_secret;
@@ -59,7 +59,7 @@ use web3::types::{Action as TraceAction, BlockId, BlockNumber, Bytes, CallReques
 use web3::{self, Web3};
 use web3_transport::{EthFeeHistoryNamespace, Web3Transport, Web3TransportNode};
 
-use super::{coin_conf, AsyncMutex, BalanceError, BalanceFut, CoinBalance, CoinFutureSpawner, CoinProtocol,
+use super::{coin_conf, AsyncMutex, BalanceError, BalanceFut, CoinBalance, CoinFutSpawner, CoinProtocol,
             CoinTransportMetrics, CoinsContext, FeeApproxStage, FoundSwapTxSpend, HistorySyncState, MarketCoinOps,
             MmCoin, MyAddressError, NegotiateSwapContractAddrErr, NumConversError, NumConversResult,
             RawTransactionError, RawTransactionFut, RawTransactionRequest, RawTransactionRes, RawTransactionResult,
@@ -315,7 +315,7 @@ pub struct EthCoinImpl {
     nonce_lock: Arc<AsyncMutex<()>>,
     erc20_tokens_infos: Arc<Mutex<HashMap<String, Erc20TokenInfo>>>,
     /// This spawner is used to spawn coin's related futures that should be aborted on coin deactivation.
-    spawner: CoinFutureSpawner,
+    spawner: AbortableSpawner,
 }
 
 #[derive(Clone, Debug)]
@@ -3078,7 +3078,7 @@ impl EthTxFeeDetails {
 impl MmCoin for EthCoin {
     fn is_asset_chain(&self) -> bool { false }
 
-    fn spawner(&self) -> &CoinFutureSpawner { &self.spawner }
+    fn spawner(&self) -> CoinFutSpawner { CoinFutSpawner::new(&self.spawner) }
 
     fn get_raw_transaction(&self, req: RawTransactionRequest) -> RawTransactionFut {
         Box::new(get_raw_transaction_impl(self.clone(), req).boxed().compat())
@@ -3656,7 +3656,7 @@ pub async fn eth_coin_from_conf_and_request(
         logs_block_range: conf["logs_block_range"].as_u64().unwrap_or(DEFAULT_LOGS_BLOCK_RANGE),
         nonce_lock,
         erc20_tokens_infos: Default::default(),
-        spawner: CoinFutureSpawner::new(),
+        spawner: AbortableSpawner::new(),
     };
     Ok(EthCoin(Arc::new(coin)))
 }
