+++ mm2src/coins/utxo/bch.rs
@@ -1144,7 +1144,7 @@ impl MarketCoinOps for BchCoin {
 impl MmCoin for BchCoin {
     fn is_asset_chain(&self) -> bool { utxo_common::is_asset_chain(&self.utxo_arc) }
 
-    fn spawner(&self) -> &CoinFutureSpawner { &self.as_ref().spawner }
+    fn spawner(&self) -> CoinFutSpawner { CoinFutSpawner::new(&self.as_ref().spawner) }
 
     fn get_raw_transaction(&self, req: RawTransactionRequest) -> RawTransactionFut {
         Box::new(utxo_common::get_raw_transaction(&self.utxo_arc, req).boxed().compat())
