+++ mm2src/coins/utxo/slp.rs
@@ -13,7 +13,7 @@ use crate::utxo::utxo_common::{self, big_decimal_from_sat_unsigned, payment_scri
 use crate::utxo::{generate_and_send_tx, sat_from_big_decimal, ActualTxFee, AdditionalTxData, BroadcastTxErr,
                   FeePolicy, GenerateTxError, RecentlySpentOutPointsGuard, UtxoCoinConf, UtxoCoinFields,
                   UtxoCommonOps, UtxoTx, UtxoTxBroadcastOps, UtxoTxGenerationOps};
-use crate::{BalanceFut, CoinBalance, CoinFutureSpawner, FeeApproxStage, FoundSwapTxSpend, HistorySyncState,
+use crate::{BalanceFut, CoinBalance, CoinFutSpawner, FeeApproxStage, FoundSwapTxSpend, HistorySyncState,
             MarketCoinOps, MmCoin, NegotiateSwapContractAddrErr, NumConversError, PrivKeyNotAllowed,
             RawTransactionFut, RawTransactionRequest, SearchForSwapTxSpendInput, SignatureResult, SwapOps, TradeFee,
             TradePreimageError, TradePreimageFut, TradePreimageResult, TradePreimageValue, TransactionDetails,
@@ -24,6 +24,7 @@ use async_trait::async_trait;
 use bitcrypto::dhash160;
 use chain::constants::SEQUENCE_FINAL;
 use chain::{OutPoint, TransactionOutput};
+use common::executor::AbortableSpawner;
 use common::log::warn;
 use common::now_ms;
 use derive_more::Display;
@@ -64,7 +65,7 @@ pub struct SlpTokenFields {
     token_id: H256,
     required_confirmations: AtomicU64,
     /// This spawner is used to spawn coin's related futures that should be aborted on coin deactivation.
-    spawner: CoinFutureSpawner,
+    spawner: AbortableSpawner,
 }
 
 /// Minimalistic info that is used to be stored outside of the token's context
@@ -280,7 +281,7 @@ impl SlpToken {
             ticker,
             token_id,
             required_confirmations: AtomicU64::new(required_confirmations),
-            spawner: CoinFutureSpawner::new(),
+            spawner: AbortableSpawner::new(),
         });
         SlpToken { conf, platform_coin }
     }
@@ -1489,7 +1490,7 @@ impl From<SlpFeeDetails> for TxFeeDetails {
 impl MmCoin for SlpToken {
     fn is_asset_chain(&self) -> bool { false }
 
-    fn spawner(&self) -> &CoinFutureSpawner { &self.conf.spawner }
+    fn spawner(&self) -> CoinFutSpawner { CoinFutSpawner::new(&self.conf.spawner) }
 
     fn get_raw_transaction(&self, req: RawTransactionRequest) -> RawTransactionFut {
         Box::new(
