+++ mm2src/coins/utxo/utxo_common_tests.rs
@@ -135,7 +135,7 @@ pub(super) fn utxo_coin_fields_for_test(
         check_utxo_maturity: false,
         block_headers_status_notifier: None,
         block_headers_status_watcher: None,
-        spawner: CoinFutureSpawner::new(),
+        spawner: AbortableSpawner::new(),
     }
 }
