+++ mm2src/coins_activation/src/bch_with_tokens_activation.rs
@@ -9,7 +9,7 @@ use coins::utxo::slp::{SlpProtocolConf, SlpToken};
 use coins::utxo::utxo_tx_history_v2::bch_and_slp_history_loop;
 use coins::utxo::UtxoCommonOps;
 use coins::{CoinBalance, CoinProtocol, MarketCoinOps, MmCoin, PrivKeyNotAllowed, UnexpectedDerivationMethod};
-use common::executor::SpawnSettings;
+use common::executor::{SpawnAbortable, SpawnSettings};
 use common::Future01CompatExt;
 use mm2_core::mm_ctx::MmArc;
 use mm2_err_handle::prelude::*;
