+++ mm2src/mm2_main/src/lp_init/init_hw.rs
@@ -163,7 +163,7 @@ impl RpcTask for InitHwTask {
 
 pub async fn init_trezor(ctx: MmArc, req: InitHwRequest) -> MmResult<InitRpcTaskResponse, InitHwError> {
     let init_ctx = MmInitContext::from_ctx(&ctx).map_to_mm(InitHwError::Internal)?;
-    let spawner = ctx.spawner.clone();
+    let spawner = ctx.spawner();
     let task = InitHwTask {
         ctx,
         hw_wallet_type: HwWalletType::Trezor,
