+++ mm2src/rpc_task/src/manager.rs
@@ -1,7 +1,7 @@
 use crate::task::RpcTaskTypes;
 use crate::{AtomicTaskId, RpcTask, RpcTaskError, RpcTaskHandle, RpcTaskResult, RpcTaskStatus, RpcTaskStatusAlias,
             TaskAbortHandle, TaskAbortHandler, TaskId, TaskStatus, TaskStatusError, UserActionSender};
-use common::executor::FutureSpawner;
+use common::executor::SpawnFuture;
 use common::log::{debug, info, warn};
 use futures::channel::oneshot;
 use futures::future::{select, Either};
@@ -31,7 +31,7 @@ impl<Task: RpcTask> RpcTaskManager<Task> {
     /// This function registers corresponding RPC task in the `RpcTaskManager` and returns the task id.
     pub fn spawn_rpc_task<F>(this: &RpcTaskManagerShared<Task>, spawner: &F, mut task: Task) -> RpcTaskResult<TaskId>
     where
-        F: FutureSpawner,
+        F: SpawnFuture,
     {
         let initial_task_status = task.initial_status();
         let (task_id, task_abort_handler) = {
